Do the File move aspects of the netcoreshim work
