+++ source/Calamari.Azure/AppServices/ConnectionStringsSetting.cs
@@ -1,23 +1,21 @@
 ﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-using Newtonsoft.Json;
+using Azure.ResourceManager.AppService.Models;
 
-namespace Calamari.AzureAppService.Json
+namespace Calamari.Azure.AppServices
 {
-    public class AppSetting
+    public class ConnectionStringSetting
     {
         public string Name { get; set; }
         public string Value { get; set; }
+        public ConnectionStringType Type { get; set; }
 
         public bool SlotSetting { get; set; }
 
-        internal void Deconstruct(out string name, out string value, out bool isSlotSetting)
+        internal void Deconstruct(out string name, out string value, out ConnectionStringType type, out bool isSlotSetting)
         {
             name = Name;
             value = Value;
+            type = Type;
             isSlotSetting = SlotSetting;
         }
 
@@ -32,7 +30,7 @@ namespace Calamari.AzureAppService.Json
         {
             if (ReferenceEquals(null, obj)) return false;
             if (ReferenceEquals(this, obj)) return true;
-            return obj.GetType() == this.GetType() && Equals((AppSetting) obj);
+            return obj.GetType() == this.GetType() && Equals((AppSetting)obj);
         }
 
         public override int GetHashCode()
@@ -41,13 +39,14 @@ namespace Calamari.AzureAppService.Json
             {
                 Name,
                 Value,
-                SlotSetting
+                SlotSetting,
+                Type
             }.GetHashCode();
         }
 
         public override string ToString()
         {
-            return $"\nName: {Name}\nValue: {Value}\nIsSlotSetting: {SlotSetting}\n";
+            return $"\nName: {Name}\nValue: {Value}\nType: {Type}\nIsSlotSetting: {SlotSetting}\n";
         }
     }
 }
