+++ source/Calamari.Azure/Calamari.Azure.csproj
@@ -12,6 +12,8 @@
     <ItemGroup>
         <PackageReference Include="Azure.Identity" Version="1.13.1" />
         <PackageReference Include="Azure.ResourceManager" Version="1.11.0" />
+        <PackageReference Include="Azure.ResourceManager.AppService" Version="1.0.2" />
+        
     </ItemGroup>
     <ItemGroup>
       <ProjectReference Include="..\Calamari.CloudAccounts\Calamari.CloudAccounts.csproj" />
