+++ source/Calamari.AzureAppService.Tests/AppServiceIntegrationTest.cs
@@ -12,6 +12,7 @@ using Azure.ResourceManager.AppService;
 using Azure.ResourceManager.AppService.Models;
 using Azure.ResourceManager.Resources;
 using Calamari.Azure;
+using Calamari.Azure.AppServices;
 using Calamari.AzureAppService.Azure;
 using Calamari.AzureAppService.Json;
 using Calamari.CloudAccounts;
