+++ source/Calamari.AzureAppService.Tests/AppServiceSettingsBehaviourFixture.cs
@@ -5,6 +5,7 @@ using System.Threading.Tasks;
 using Azure;
 using Azure.ResourceManager.AppService.Models;
 using Azure.ResourceManager.Resources;
+using Calamari.Azure.AppServices;
 using Calamari.AzureAppService.Azure;
 using Calamari.AzureAppService.Behaviors;
 using Calamari.AzureAppService.Json;
