+++ source/Calamari.AzureAppService/Behaviors/AzureAppServiceAppSettingsBehaviour.cs
@@ -7,11 +7,10 @@ using System.Threading.Tasks;
 using Azure.ResourceManager;
 using Azure.ResourceManager.AppService.Models;
 using Calamari.Azure;
+using Calamari.Azure.AppServices;
 using Calamari.AzureAppService.Azure;
-using Calamari.AzureAppService.Json;
 using Calamari.CloudAccounts;
 using Calamari.Common.Commands;
-using Calamari.Common.FeatureToggles;
 using Calamari.Common.Plumbing.Logging;
 using Calamari.Common.Plumbing.Pipeline;
 using Newtonsoft.Json;
